trace-agent-bootstrap/src/main/java/com/vmlens/trace/agent/bootstrap/callback/callbackaction/CallbackActionEndAtomicOperation.java
package com.vmlens.trace.agent.bootstrap.callback.callbackaction;

import com.vmlens.trace.agent.bootstrap.callback.threadlocal.ThreadLocalWhenInTest;
import com.vmlens.trace.agent.bootstrap.event.RuntimeEvent;
import com.vmlens.trace.agent.bootstrap.event.SerializableEvent;
import com.vmlens.trace.agent.bootstrap.util.TLinkableWrapper;
import gnu.trove.list.linked.TLinkedList;

import static com.vmlens.trace.agent.bootstrap.util.TLinkableWrapper.emptyList;

public class CallbackActionEndAtomicOperation implements CallbackAction {
    @Override
    public TLinkedList<TLinkableWrapper<SerializableEvent>> execute(ThreadLocalWhenInTest threadLocalDataWhenInTest) {
        RuntimeEvent event = threadLocalDataWhenInTest.atomicOperation().create();
        threadLocalDataWhenInTest.after(event);
        threadLocalDataWhenInTest.setAtomicOperation(null);
        return emptyList();
    }
}
